+++ packages/snap/jest.config.js
@@ -24,8 +24,8 @@ const config = {
   // An object that configures minimum threshold enforcement for coverage results
   coverageThreshold: {
     global: {
-      branches: 62.1,
-      functions: 67.28,
+      branches: 62.15,
+      functions: 67.4,
       lines: 76.86,
       statements: 76.86,
     },